performing below actions to generate the ascii art from image.
Get an image path as a command-line argument.
Open the image from the provided path.
Calculate the aspect ratio.
Resize the image. Here we are taking the new width as 120 pixels. Adjusting the new height as per aspect ratio.
Convert the image into a greyscale format.
Get all the pixels of the image. Replace pixels with intensity in a defined range with a character from the list.
Print image and save to text file.

If you are getting an elongated image in the terminal, you can adjust the height accordingly. I am generating new_height using formula new_height = aspect_ratio * new_width * 0.55.

Statement list(img.getdata()) returns the list intensity of all pixels. The value of intensity will be between 0 to 255. Lower the value, the darker the color. That is why we have characters in the char list arranged from higher intensity to lower intensity.

Value of pixel//25, where 25 is the intensity range for one character, can be between 0 to 11 i.e why there are 11 characters in charslist. You may keep more characters in the list, let's say 16 characters, and then keep the range size as 17.

Line new_pixels = [chars[pixel//25] for the pixel in pixels] replaces the pixels with intensity from 0 to 25 with the first character from the chars list and 26 to 50 with seconds character from chars list and so on.